refactor(apple): move NavigationStack under TabView, move Routers under NavigationStacks, refactor root views
